// src/lib/cv/types.ts
export interface Language {
  id?: string;
  name?: string;
  language?: string;
  level?: string;
  proficiency?: string;
  certification_name?: string;
}